change file structure; add ssg for disc test; add shared lib components usage;
