+++ apps/_compability_app/containers/GlobalErrorBoundery.jsx
