+++ apps/_compability_app/pages/auth/register.jsx
@@ -15,7 +15,7 @@ import gql from 'graphql-tag'
 
 import Router from 'next/router'
 import { BaseLayout } from '../../containers'
-import { useAuth } from '../../lib/auth'
+import { useAuth } from '@core/next/auth'
 import { useIntl } from 'react-intl'
 
 const { Title } = Typography
